// cx-competence/src/app/shared/constants/survey-variable.enum.ts
export enum SurveyVariableEnum {
  formDisplayMode = 'formDisplayMode',
  currentUser_id = 'currentUser_id', // We are not using this yet because CXID doesn't return this info.
  currentUser_extId = 'currentUser_extId',
  currentUser_systemRoles = 'currentUser_systemRoles',
  currentUser_assignRolePermission = 'currentUser_assignRolePermission', // This is only applicable for SAM.
  currentDepartment_id = 'currentDepartment_id',
  currentDepartment_types = 'currentDepartment_types',
  currentDepartment_name = 'currentDepartment_name',
  currentDepartment_archetype = 'currentDepartment_archetype',
  currentObject_id = 'currentObject_id',
  currentObject_extId = 'currentObject_extId',
  currentObject_isExternallyMastered = 'currentObject_isExternallyMastered',
  currentObject_entityStatus = 'currentObject_entityStatus',
  currentObject_statusType = 'currentObject_statusType',
  currentObject_departmentId = 'currentObject_departmentId',
  currentObject_departmentTypes = 'currentObject_departmentTypes',
  currentObject_departmentName = 'currentObject_departmentName',
  currentObject_departmentArchetype = 'currentObject_departmentArchetype',
  currentObject_userTypes = 'currentObject_userTypes',
  currentObject_userServiceSchemes = 'currentObject_userServiceSchemes',
  currentObject_userDevelopmentalRoles = 'currentObject_userDevelopmentalRoles',
  currentObject_userLearningFrameworks = 'currentObject_userLearningFrameworks',
  competenceApi_BaseUrl = 'competenceApi_BaseUrl',
  organizationApi_BaseUrl = 'organizationApi_BaseUrl',
  assessmentApi_BaseUrl = 'assessmentApi_BaseUrl',
  learningcatalogApi_BaseUrl = 'learningcatalogApi_BaseUrl',
  cxCategoryTypeId = 'cxCategoryTypeId',
  surveyTimeStamp = 'surveyTimeStamp',
  displayApproval = 'displayApproval',
}